+++ src/Apps/MovementPathCli/main.cpp
@@ -119,7 +119,6 @@ int main()
 
     VoxelPathPlannerOptions fastOptions =
         VoxelPathPlanner::MakeDefaultOptions();
-    fastOptions.localBuildOptions.regionMode = VoxelBuildRegionMode::StartGoalBox;
     fastOptions.runOptions.exportVtk = true;
     fastOptions.runOptions.debugNeighborhood = false;
     fastOptions.runOptions.verbose = true;